+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version Alpha 0.0.13a
 * Changed application name from "WifftOCR" to "OCR Studio".
 * Make app icon to be displayed in title bar instead of just red circle.
+* Temp files generated during recording now are cleaned up when OCR recorder service stops.
 
 ## Version Alpha 0.0.12a_03 (2023-05-19) [PUBLIC]
 * Fixed log file not being locked while reading.